Don't just print stacktrace with precondition parse errors
