+++ liquibase-core/src/main/java/liquibase/precondition/PreconditionLogic.java
@@ -3,6 +3,7 @@ package liquibase.precondition;
 import liquibase.GlobalConfiguration;
 import liquibase.database.Database;
 import liquibase.exception.ValidationErrors;
+import liquibase.parser.ChangeLogParserConfiguration;
 import liquibase.parser.core.ParsedNode;
 import liquibase.parser.core.ParsedNodeException;
 import liquibase.resource.ResourceAccessor;
@@ -48,7 +49,7 @@ public abstract class PreconditionLogic extends AbstractPrecondition {
     protected Precondition toPrecondition(ParsedNode node, ResourceAccessor resourceAccessor) throws ParsedNodeException {
         Precondition precondition = PreconditionFactory.getInstance().create(node.getName());
         if (precondition == null) {
-            if (node.getChildren() != null && node.getChildren().size() > 0 && GlobalConfiguration.CHANGELOG_PARSE_MODE.getCurrentValue().equals(GlobalConfiguration.ChangelogParseMode.STRICT)) {
+            if (node.getChildren() != null && node.getChildren().size() > 0 && ChangeLogParserConfiguration.CHANGELOG_PARSE_MODE.getCurrentValue().equals(ChangeLogParserConfiguration.ChangelogParseMode.STRICT)) {
                 throw new ParsedNodeException("Unknown precondition '" + node.getName() + "'. Check for spelling or capitalization errors and missing extensions such as liquibase-commercial.");
             }
